refactor(npc-tool): split main.rs CLI into modules + drop validate macro (#1200)
- TD-027: replace validate_db's with_filter inner fn + count! macro with a
  count_with_optional_parish() free function; each rule is now a one-liner.
- TD-028: decompose the 2,050-LOC single-file CLI (main.rs) into focused
  modules: db, schema, generate, query, validate, import_export, tests.
  main.rs keeps clap wiring, main(), and the shared DataTier/ExportNpc/
  ExportBlob types. Command behavior is unchanged; 59 tests pass.
