+++ parish/crates/parish-npc-tool/src/main.rs
@@ -1,86 +1,33 @@
 mod catalog;
+mod db;
+mod generate;
+mod import_export;
+mod query;
+mod schema;
+mod validate;
 
-use std::io::{self, Read};
-use std::path::{Path, PathBuf};
+// Bring the split command implementations into the crate root so `main()`
+// and the `#[cfg(test)] mod tests` (via `use super::*`) keep their existing
+// unqualified call sites after the TD-028 module split.
+use db::{open_db, resolve_default_db};
+use generate::{elaborate_parish, generate_parish, generate_world};
+use import_export::{export_npcs, import_npcs};
+use query::{
+    edit_npc, family_tree, list_npcs, promote_npc, relationships, search_npcs, show_npc, stats,
+};
+use schema::ensure_schema;
+use validate::validate_db;
 
-use anyhow::{Context, Result, bail};
+use std::path::PathBuf;
+
+use anyhow::Result;
 use clap::{Parser, Subcommand, ValueEnum};
-use rand::prelude::IndexedRandom;
-use rand::{Rng, SeedableRng, rngs::StdRng};
-use rusqlite::{Connection, OptionalExtension, params};
 use serde::{Deserialize, Serialize};
-
-/// Leaf filename of the NPC-tool database, relative to the `data/` directory.
-const DB_FILENAME: &str = "parish-world.db";
-
-/// Environment variable that pins the DB path directly (overrides project-root
-/// anchor but is overridden by an explicit `--db` flag on the command line).
-const NPC_TOOL_DB_ENV: &str = "PARISH_NPC_TOOL_DB";
 
 /// Current world year for the simulation. Used to derive `birth_year` from an
 /// NPC's age in both `generate_parish` and `import_npcs_inner`; keeping it in
 /// one place prevents the two call sites from silently diverging.
-const WORLD_YEAR: i64 = 1820;
-
-/// Resolves the default NPC-tool DB path when `--db` is not given on the
-/// command line.  Resolution order (Rule 9 — never bare cwd-relative):
-///
-///  1. `PARISH_NPC_TOOL_DB` env var — explicit operator/test override.
-///  2. `PARISH_DATA_DIR` env var — mirrors the data-dir convention used by
-///     the server and Tauri entry-points; appends `parish-world.db`.
-///  3. Walk up to 4 ancestor directories of the startup cwd looking for
-///     `Cargo.toml` (project root sentinel); returns
-///     `<root>/data/parish-world.db` when found.
-///  4. Bare `data/parish-world.db` relative to the startup cwd as a last
-///     resort — matches the previous hard-coded default so single-shot runs
-///     from the repo root still work.
-pub fn resolve_default_db() -> PathBuf {
-    // 1. Explicit env-var override.
-    if let Ok(s) = std::env::var(NPC_TOOL_DB_ENV) {
-        let trimmed = s.trim();
-        if !trimmed.is_empty() {
-            return PathBuf::from(trimmed);
-        }
-    }
-
-    // 2. PARISH_DATA_DIR mirrors the convention in parish-tauri / parish-server.
-    if let Ok(data_dir) = std::env::var("PARISH_DATA_DIR") {
-        let trimmed = data_dir.trim();
-        if !trimmed.is_empty() {
-            return PathBuf::from(trimmed).join(DB_FILENAME);
-        }
-    }
-
-    // 3. Walk ancestors for Cargo.toml (project root).
-    let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-    let mut p = cwd.clone();
-    for _ in 0..4 {
-        if p.join("Cargo.toml").exists() {
-            return p.join("data").join(DB_FILENAME);
-        }
-        match p.parent() {
-            Some(parent) => p = parent.to_path_buf(),
-            None => break,
-        }
-    }
-
-    // 4. Last resort: startup-cwd relative (prior behaviour).
-    cwd.join("data").join(DB_FILENAME)
-}
-
-const MALE_NAMES: &[&str] = &["Pádraig", "Seán", "Michael", "Thomas", "James", "Brendan"];
-const FEMALE_NAMES: &[&str] = &["Mary", "Bridget", "Margaret", "Catherine", "Niamh", "Aoife"];
-const SURNAMES: &[&str] = &["Kelly", "Murphy", "Brennan", "O'Brien", "Flanagan", "Darcy"];
-const OCCUPATIONS: &[(&str, u8)] = &[
-    ("Tenant Farmer", 35),
-    ("Laborer", 30),
-    ("Servant", 10),
-    ("Craftsman", 8),
-    ("Shopkeeper", 3),
-    ("Clergy", 1),
-    ("Gentry", 2),
-    ("Other", 11),
-];
+pub(crate) const WORLD_YEAR: i64 = 1820;
 
 #[derive(Parser, Debug)]
 #[command(name = "parish-npc-tool")]
@@ -189,7 +136,7 @@ enum Command {
 }
 
 #[derive(Debug, Clone, Copy, ValueEnum)]
-enum DataTier {
+pub(crate) enum DataTier {
     Sketched,
     Elaborated,
     Authored,
@@ -215,12 +162,12 @@ impl DataTier {
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-struct ExportBlob {
+pub(crate) struct ExportBlob {
     npcs: Vec<ExportNpc>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-struct ExportNpc {
+pub(crate) struct ExportNpc {
     id: i64,
     name: String,
     /// NPC sex ("female" | "male" | "unknown"). Added in #436 so
@@ -237,7 +184,7 @@ struct ExportNpc {
     personality: Option<String>,
 }
 
-fn default_sex() -> String {
+pub(crate) fn default_sex() -> String {
     "unknown".to_string()
 }
 
@@ -299,1752 +246,5 @@ fn main() -> Result<()> {
     }
 }
 
-fn open_db(path: &Path) -> Result<Connection> {
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent).with_context(|| {
-            format!("failed to create DB parent directory {}", parent.display())
-        })?;
-    }
-    let conn =
-        Connection::open(path).with_context(|| format!("failed to open {}", path.display()))?;
-    ensure_schema(&conn)?;
-    Ok(conn)
-}
-
-/// Initialises the standalone `parish-npc-tool` SQLite schema.
-///
-/// # Schema divergence from parish-persistence (#434)
-///
-/// This schema is **not** compatible with the main game's persistence
-/// format in `parish-persistence` (which stores branch-based game
-/// snapshots keyed by session id, not relational parish/household/NPC
-/// rows). Databases created by `parish-npc-tool` cannot be loaded by the
-/// running game, and save files created by the game cannot be opened
-/// by `parish-npc-tool` commands.
-///
-/// That divergence is deliberate: `parish-npc-tool` is a world-*building*
-/// tool that authors use at design time to generate large populations
-/// with relational constraints (households, relationships, validation
-/// sweeps). The runtime engine only needs read-only NPC data and
-/// materialises it into the in-memory `NpcManager` from
-/// `mods/<name>/npcs.json`. The two codepaths have different
-/// workloads and different shape — forcing them into one schema would
-/// burden the runtime with author-time fields (`data_tier`, parish
-/// joins) it doesn't use, or would starve the CLI of the relational
-/// structure it depends on.
-///
-/// The practical round-trip is:
-///
-/// 1. `parish-npc-tool generate-parish …` populates this schema.
-/// 2. `parish-npc-tool export [--parish NAME]` emits the JSON blob the
-///    game consumes, which can be hand-massaged into `npcs.json`.
-/// 3. The game loads `npcs.json` into its own runtime structures —
-///    no direct SQLite interop.
-///
-/// If you need to hold both formats in sync, treat the parish-npc-tool DB
-/// as the source of truth at design time and re-export after every
-/// authoring session. A proper conversion utility between this schema
-/// and a gameplay save is out of scope; track additions there under
-/// #434 follow-ups.
-fn ensure_schema(conn: &Connection) -> Result<()> {
-    conn.execute_batch(
-        "
-        CREATE TABLE IF NOT EXISTS counties (
-            id INTEGER PRIMARY KEY,
-            name TEXT UNIQUE NOT NULL
-        );
-        CREATE TABLE IF NOT EXISTS parishes (
-            id INTEGER PRIMARY KEY,
-            county_id INTEGER,
-            name TEXT UNIQUE NOT NULL,
-            FOREIGN KEY(county_id) REFERENCES counties(id)
-        );
-        CREATE TABLE IF NOT EXISTS households (
-            id INTEGER PRIMARY KEY,
-            parish_id INTEGER NOT NULL,
-            name TEXT NOT NULL,
-            FOREIGN KEY(parish_id) REFERENCES parishes(id)
-        );
-        CREATE TABLE IF NOT EXISTS npcs (
-            id INTEGER PRIMARY KEY,
-            name TEXT NOT NULL,
-            sex TEXT NOT NULL,
-            birth_year INTEGER NOT NULL,
-            age INTEGER NOT NULL,
-            parish_id INTEGER NOT NULL,
-            household_id INTEGER,
-            occupation TEXT NOT NULL,
-            data_tier INTEGER NOT NULL DEFAULT 0,
-            mood TEXT,
-            personality TEXT,
-            FOREIGN KEY(parish_id) REFERENCES parishes(id),
-            FOREIGN KEY(household_id) REFERENCES households(id)
-        );
-        CREATE TABLE IF NOT EXISTS npc_relationships (
-            from_npc_id INTEGER NOT NULL,
-            to_npc_id INTEGER NOT NULL,
-            kind TEXT NOT NULL,
-            strength REAL NOT NULL,
-            PRIMARY KEY (from_npc_id, to_npc_id),
-            FOREIGN KEY(from_npc_id) REFERENCES npcs(id),
-            FOREIGN KEY(to_npc_id) REFERENCES npcs(id)
-        );
-        CREATE INDEX IF NOT EXISTS idx_npcs_parish ON npcs(parish_id);
-        CREATE INDEX IF NOT EXISTS idx_npcs_occupation ON npcs(occupation);
-        CREATE INDEX IF NOT EXISTS idx_npcs_tier ON npcs(data_tier);
-        CREATE INDEX IF NOT EXISTS idx_npcs_name ON npcs(name);
-    ",
-    )
-    .context("failed to create schema")?;
-    Ok(())
-}
-
-fn generate_world(conn: &Connection, counties: &[String]) -> Result<()> {
-    if counties.is_empty() {
-        bail!("--counties is required (comma-separated)");
-    }
-    for county in counties {
-        conn.execute(
-            "INSERT OR IGNORE INTO counties(name) VALUES (?)",
-            params![county.to_lowercase()],
-        )?;
-    }
-    println!("Seeded {} counties", counties.len());
-    Ok(())
-}
-
-fn generate_parish(conn: &Connection, parish: &str, pop: u32, seed: Option<u64>) -> Result<()> {
-    // Ensure a county row exists before opening the generation transaction
-    // so that the INSERT OR IGNORE below sees a valid county_id.
-    //
-    // TD-023: never auto-create a hard-coded county (it surprised authors of
-    // mods set in Galway/Mayo) and never `.expect()`-panic on a read-only or
-    // constrained DB. Require the caller to seed counties via `generate-world`
-    // first, surfacing a clear recoverable error instead of a process abort.
-    let county_id: i64 = match conn
-        .query_row("SELECT id FROM counties ORDER BY id LIMIT 1", [], |r| {
-            r.get(0)
-        })
-        .optional()?
-    {
-        Some(id) => id,
-        None => bail!(
-            "no county exists — run `generate-world --counties <name,...>` before `generate-parish`"
-        ),
-    };
-
-    // Wrap all NPC/relationship inserts in a single transaction (#606).
-    // If any insert fails (disk full, constraint violation, process crash)
-    // the entire generation is rolled back, leaving no orphaned rows.
-    // Matches the transaction pattern used by `import_npcs_inner`.
-    let tx = conn.unchecked_transaction()?;
-
-    let parish_lc = parish.to_lowercase();
-    tx.execute(
-        "INSERT OR IGNORE INTO parishes(county_id, name) VALUES (?, ?)",
-        params![county_id, &parish_lc],
-    )?;
-    let parish_id: i64 = tx.query_row(
-        "SELECT id FROM parishes WHERE name = ?",
-        params![&parish_lc],
-        |r| r.get(0),
-    )?;
-
-    let mut rng = StdRng::seed_from_u64(seed.unwrap_or(42));
-    let household_count = (pop / 6).max(1);
-    let now_year = WORLD_YEAR;
-
-    for i in 0..household_count {
-        tx.execute(
-            "INSERT INTO households(parish_id, name) VALUES (?, ?)",
-            params![parish_id, format!("{} Household {}", parish, i + 1)],
-        )?;
-        let household_id = tx.last_insert_rowid();
-        let members = rng.random_range(4..=8);
-        for _ in 0..members {
-            let female = rng.random_bool(0.5);
-            let first = if female {
-                FEMALE_NAMES
-                    .choose(&mut rng)
-                    .expect("female names list is non-empty")
-            } else {
-                MALE_NAMES
-                    .choose(&mut rng)
-                    .expect("male names list is non-empty")
-            };
-            let surname = SURNAMES
-                .choose(&mut rng)
-                .expect("surname list is non-empty");
-            let name = format!("{} {}", first, surname);
-            let age: i64 = rng.random_range(0..=85);
-            let birth_year = now_year - age;
-            let occupation = weighted_occupation(&mut rng);
-            tx.execute(
-                "INSERT INTO npcs(name, sex, birth_year, age, parish_id, household_id, occupation, data_tier, mood) VALUES (?, ?, ?, ?, ?, ?, ?, 0, ?)",
-                params![name, if female {"female"} else {"male"}, birth_year, age, parish_id, household_id, occupation, "neutral"],
-            )?;
-        }
-    }
-
-    let mut stmt = tx.prepare("SELECT id FROM npcs WHERE parish_id = ?")?;
-    let npc_ids: Vec<i64> = stmt
-        .query_map(params![parish_id], |r| r.get(0))?
-        .collect::<rusqlite::Result<Vec<_>>>()?;
-    drop(stmt);
-
-    let mut relationships: Vec<(i64, i64, f64)> = Vec::new();
-    for id in &npc_ids {
-        for _ in 0..2 {
-            if let Some(other) = npc_ids.choose(&mut rng)
-                && other != id
-            {
-                relationships.push((*id, *other, rng.random_range(-0.2..0.9)));
-            }
-        }
-    }
-    for (from, to, strength) in &relationships {
-        tx.execute(
-            "INSERT OR IGNORE INTO npc_relationships(from_npc_id, to_npc_id, kind, strength) VALUES (?, ?, ?, ?)",
-            params![from, to, "Acquaintance", strength],
-        )?;
-    }
-
-    let count: i64 = tx.query_row(
-        "SELECT COUNT(*) FROM npcs WHERE parish_id = ?",
-        params![parish_id],
-        |r| r.get(0),
-    )?;
-    tx.commit()?;
-    println!("Generated parish '{}' with {} sketched NPCs", parish, count);
-    Ok(())
-}
-
-fn weighted_occupation(rng: &mut StdRng) -> &'static str {
-    let roll: u8 = rng.random_range(0..100);
-    let mut acc = 0_u8;
-    for (occ, weight) in OCCUPATIONS {
-        acc = acc.saturating_add(*weight);
-        if roll < acc {
-            return occ;
-        }
-    }
-    "Other"
-}
-
-fn list_npcs(
-    conn: &Connection,
-    parish: Option<&str>,
-    occupation: Option<&str>,
-    tier: Option<DataTier>,
-    limit: u32,
-) -> Result<()> {
-    let mut clauses: Vec<String> = Vec::new();
-    let mut bind: Vec<String> = Vec::new();
-
-    if let Some(p) = parish {
-        clauses.push("p.name = ?".to_string());
-        bind.push(p.to_lowercase());
-    }
-    if let Some(o) = occupation {
-        clauses.push("n.occupation = ?".to_string());
-        bind.push(o.to_string());
-    }
-    if let Some(t) = tier {
-        clauses.push("n.data_tier = ?".to_string());
-        bind.push(t.as_i64().to_string());
-    }
-
-    let where_clause = if clauses.is_empty() {
-        String::new()
-    } else {
-        format!("WHERE {} ", clauses.join(" AND "))
-    };
-    let sql = format!(
-        "SELECT n.id, n.name, p.name, n.occupation, n.data_tier \
-         FROM npcs n JOIN parishes p ON p.id = n.parish_id \
-         {}ORDER BY n.id LIMIT ?",
-        where_clause
-    );
-    bind.push(limit.to_string());
-
-    let mut stmt = conn.prepare(&sql)?;
-    let mut rows = stmt.query(rusqlite::params_from_iter(bind.iter()))?;
-    println!("id\tname\tparish\toccupation\ttier");
-    while let Some(row) = rows.next()? {
-        let tier_i: i64 = row.get(4)?;
-        println!(
-            "{}\t{}\t{}\t{}\t{}",
-            row.get::<_, i64>(0)?,
-            row.get::<_, String>(1)?,
-            row.get::<_, String>(2)?,
-            row.get::<_, String>(3)?,
-            DataTier::from_i64(tier_i)
-        );
-    }
-    Ok(())
-}
-
-fn show_npc(conn: &Connection, npc_id: i64) -> Result<()> {
-    let row = conn
-        .query_row(
-            "
-            SELECT n.id, n.name, n.age, p.name, n.occupation, n.data_tier, n.mood, n.personality
-            FROM npcs n JOIN parishes p ON p.id = n.parish_id
-            WHERE n.id = ?
-            ",
-            params![npc_id],
-            |r| {
-                Ok((
-                    r.get::<_, i64>(0)?,
-                    r.get::<_, String>(1)?,
-                    r.get::<_, i64>(2)?,
-                    r.get::<_, String>(3)?,
-                    r.get::<_, String>(4)?,
-                    r.get::<_, i64>(5)?,
-                    r.get::<_, Option<String>>(6)?,
-                    r.get::<_, Option<String>>(7)?,
-                ))
-            },
-        )
-        .optional()?;
-
-    if let Some((id, name, age, parish, occupation, tier, mood, personality)) = row {
-        println!("id: {id}");
-        println!("name: {name}");
-        println!("age: {age}");
-        println!("parish: {parish}");
-        println!("occupation: {occupation}");
-        println!("tier: {}", DataTier::from_i64(tier));
-        println!("mood: {}", mood.unwrap_or_else(|| "-".to_string()));
-        println!(
-            "personality: {}",
-            personality.unwrap_or_else(|| "(none)".to_string())
-        );
-        Ok(())
-    } else {
-        bail!("NPC {} not found", npc_id)
-    }
-}
-
-fn escape_like(s: &str) -> String {
-    s.replace('\\', "\\\\")
-        .replace('%', "\\%")
-        .replace('_', "\\_")
-}
-
-fn search_npcs(conn: &Connection, query: &str, limit: u32) -> Result<()> {
-    let like = format!("%{}%", escape_like(query));
-    let mut stmt = conn.prepare(
-        "
-        SELECT n.id, n.name, p.name, n.occupation
-        FROM npcs n JOIN parishes p ON p.id = n.parish_id
-        WHERE n.name LIKE ? ESCAPE '\\'
-        ORDER BY n.name
-        LIMIT ?
-    ",
-    )?;
-    let rows = stmt.query_map(params![like, limit], |r| {
-        Ok((
-            r.get::<_, i64>(0)?,
-            r.get::<_, String>(1)?,
-            r.get::<_, String>(2)?,
-            r.get::<_, String>(3)?,
-        ))
-    })?;
-
-    for row in rows {
-        let (id, name, parish, occupation) = row?;
-        println!("{id}: {name} ({occupation}, {parish})");
-    }
-    Ok(())
-}
-
-fn edit_npc(
-    conn: &Connection,
-    npc_id: i64,
-    mood: Option<String>,
-    occupation: Option<String>,
-) -> Result<()> {
-    if mood.is_none() && occupation.is_none() {
-        bail!("provide at least one change (--mood or --occupation)");
-    }
-    if let Some(m) = mood {
-        conn.execute("UPDATE npcs SET mood = ? WHERE id = ?", params![m, npc_id])?;
-    }
-    if let Some(o) = occupation {
-        conn.execute(
-            "UPDATE npcs SET occupation = ? WHERE id = ?",
-            params![o, npc_id],
-        )?;
-    }
-    println!("Updated NPC {}", npc_id);
-    Ok(())
-}
-
-fn promote_npc(conn: &Connection, npc_id: i64) -> Result<()> {
-    let changed = conn.execute(
-        "
-        UPDATE npcs
-        SET data_tier = 1,
-            personality = COALESCE(personality, 'A quietly observant parishioner with strong local ties.'),
-            mood = COALESCE(mood, 'curious')
-        WHERE id = ?
-    ",
-        params![npc_id],
-    )?;
-    if changed == 0 {
-        bail!("NPC {} not found", npc_id);
-    }
-    println!("Promoted NPC {} to Elaborated", npc_id);
-    Ok(())
-}
-
-fn elaborate_parish(conn: &Connection, parish: &str, batch: u32) -> Result<()> {
-    let parish = parish.to_lowercase();
-    let mut stmt = conn.prepare(
-        "
-        SELECT n.id
-        FROM npcs n JOIN parishes p ON p.id = n.parish_id
-        WHERE p.name = ? AND n.data_tier = 0
-        ORDER BY n.id
-        LIMIT ?
-    ",
-    )?;
-    let ids: Vec<i64> = stmt
-        .query_map(params![parish, batch], |r| r.get(0))?
-        .collect::<rusqlite::Result<Vec<_>>>()?;
-
-    for id in &ids {
-        promote_npc(conn, *id)?;
-    }
-    println!("Elaborated {} NPCs in parish {}", ids.len(), parish);
-    Ok(())
-}
-
-fn validate_db(conn: &Connection, parish: Option<String>, all: bool) -> Result<()> {
-    if parish.is_some() && all {
-        bail!("choose either --parish or --all");
-    }
-
-    // #592 — Use a parameterized sub-query for the parish filter so the
-    // caller-supplied name is never interpolated into SQL text.  The parish
-    // name is passed as a bound value to `rusqlite`; the predicate columns
-    // (`household_id`, `age`, …) are fixed literals that never originate from
-    // user input, so they remain safe to format into the SQL string.
-    let has_parish = parish.is_some();
-    let parish_name: String = parish.as_deref().unwrap_or("").to_lowercase();
-
-    // Build a counting query for a fixed predicate.  When `has_parish` is
-    // true the query adds a second `?` placeholder for the parish name and
-    // the caller must supply it as the last element of the params tuple.
-    fn with_filter(predicate: &str, has_parish: bool) -> String {
-        if has_parish {
-            format!(
-                "SELECT COUNT(*) FROM npcs n \
-                 WHERE n.parish_id = (SELECT id FROM parishes WHERE name = ?) \
-                 AND ({predicate})"
-            )
-        } else {
-            format!("SELECT COUNT(*) FROM npcs n WHERE {predicate}")
-        }
-    }
-
-    macro_rules! count {
-        ($predicate:expr) => {{
-            let sql = with_filter($predicate, has_parish);
-            if has_parish {
-                conn.query_row(&sql, params![&parish_name], |r| r.get::<_, i64>(0))?
-            } else {
-                conn.query_row(&sql, [], |r| r.get::<_, i64>(0))?
-            }
-        }};
-    }
-
-    let missing_households: i64 =
-        count!("household_id IS NULL OR household_id NOT IN (SELECT id FROM households)");
-    let invalid_age: i64 = count!("age < 0 OR age > 110");
-    let elaborated_without_personality: i64 =
-        count!("data_tier >= 1 AND (personality IS NULL OR personality = '')");
-    let broken_relationships: i64 = conn.query_row(
-        "
-        SELECT COUNT(*) FROM npc_relationships r
-        LEFT JOIN npcs a ON a.id = r.from_npc_id
-        LEFT JOIN npcs b ON b.id = r.to_npc_id
-        WHERE a.id IS NULL OR b.id IS NULL
-    ",
-        [],
-        |r| r.get(0),
-    )?;
-
-    println!("Validation report:");
-    println!("- missing_households: {missing_households}");
-    println!("- invalid_age: {invalid_age}");
-    println!("- elaborated_without_personality: {elaborated_without_personality}");
-    println!("- broken_relationships: {broken_relationships}");
-
-    if missing_households + invalid_age + elaborated_without_personality + broken_relationships > 0
-    {
-        bail!("validation failed");
-    }
-    println!("Validation passed");
-    Ok(())
-}
-
-fn stats(conn: &Connection) -> Result<()> {
-    let total: i64 = conn.query_row("SELECT COUNT(*) FROM npcs", [], |r| r.get(0))?;
-    let sketched: i64 =
-        conn.query_row("SELECT COUNT(*) FROM npcs WHERE data_tier = 0", [], |r| {
-            r.get(0)
-        })?;
-    let elaborated: i64 =
-        conn.query_row("SELECT COUNT(*) FROM npcs WHERE data_tier = 1", [], |r| {
-            r.get(0)
-        })?;
-    let authored: i64 =
-        conn.query_row("SELECT COUNT(*) FROM npcs WHERE data_tier = 2", [], |r| {
-            r.get(0)
-        })?;
-
-    println!("Total NPCs: {total}");
-    println!("Sketched: {sketched}");
-    println!("Elaborated: {elaborated}");
-    println!("Authored: {authored}");
-    Ok(())
-}
-
-fn export_npcs(conn: &Connection, parish: Option<&str>) -> Result<()> {
-    // `sex` added in #436 so import can restore it rather than
-    // hard-coding 'unknown'. Keep the column order stable so the
-    // mapper indices are obvious.
-    let sql = if parish.is_some() {
-        "
-        SELECT n.id, n.name, n.sex, n.age, p.name, n.occupation, n.data_tier, n.mood, n.personality
-        FROM npcs n JOIN parishes p ON p.id = n.parish_id
-        WHERE p.name = ?
-        ORDER BY n.id
-        "
-    } else {
-        "
-        SELECT n.id, n.name, n.sex, n.age, p.name, n.occupation, n.data_tier, n.mood, n.personality
-        FROM npcs n JOIN parishes p ON p.id = n.parish_id
-        ORDER BY n.id
-        "
-    };
-    let mut stmt = conn.prepare(sql)?;
-
-    let mapper = |r: &rusqlite::Row<'_>| {
-        Ok(ExportNpc {
-            id: r.get(0)?,
-            name: r.get(1)?,
-            sex: r.get(2)?,
-            age: r.get(3)?,
-            parish: r.get(4)?,
-            occupation: r.get(5)?,
-            data_tier: r.get(6)?,
-            mood: r.get(7)?,
-            personality: r.get(8)?,
-        })
-    };
-    let npcs = if let Some(p) = parish {
-        stmt.query_map(params![p.to_lowercase()], mapper)?
-            .collect::<rusqlite::Result<Vec<_>>>()?
-    } else {
-        stmt.query_map([], mapper)?
-            .collect::<rusqlite::Result<Vec<_>>>()?
-    };
-
-    let blob = ExportBlob { npcs };
-    println!("{}", serde_json::to_string_pretty(&blob)?);
-    Ok(())
-}
-
-fn import_npcs_inner(conn: &Connection, npcs: Vec<ExportNpc>) -> Result<(u64, u64)> {
-    let tx = conn.unchecked_transaction()?;
-    let mut inserted = 0u64;
-    let mut updated = 0u64;
-    for npc in npcs {
-        let parish_lc = npc.parish.to_lowercase();
-        tx.execute(
-            "INSERT OR IGNORE INTO parishes(county_id, name) VALUES ((SELECT id FROM counties LIMIT 1), ?)",
-            params![&parish_lc],
-        )?;
-        let parish_id: i64 = tx.query_row(
-            "SELECT id FROM parishes WHERE name = ?",
-            params![&parish_lc],
-            |r| r.get(0),
-        )?;
-
-        // #436: use INSERT … ON CONFLICT DO UPDATE instead of INSERT
-        // OR REPLACE so columns that aren't in the export blob (most
-        // importantly `household_id`) are preserved on existing rows.
-        // INSERT OR REPLACE deletes the old row and inserts a new
-        // one, silently losing household_id, personality when the
-        // blob doesn't include it, etc. The `sex` column now comes
-        // from the blob so export→import is a lossless round-trip.
-        let row_existed_before: bool = tx
-            .query_row(
-                "SELECT EXISTS(SELECT 1 FROM npcs WHERE id = ?)",
-                params![npc.id],
-                |r| r.get(0),
-            )
-            .unwrap_or(false);
-
-        tx.execute(
-            "INSERT INTO npcs(id, name, sex, birth_year, age, parish_id, occupation, data_tier, mood, personality)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-             ON CONFLICT(id) DO UPDATE SET
-                 name        = excluded.name,
-                 sex         = excluded.sex,
-                 birth_year  = excluded.birth_year,
-                 age         = excluded.age,
-                 parish_id   = excluded.parish_id,
-                 occupation  = excluded.occupation,
-                 data_tier   = excluded.data_tier,
-                 mood        = excluded.mood,
-                 personality = excluded.personality",
-            params![
-                npc.id,
-                npc.name,
-                npc.sex,
-                WORLD_YEAR - npc.age,
-                npc.age,
-                parish_id,
-                npc.occupation,
-                npc.data_tier,
-                npc.mood,
-                npc.personality
-            ],
-        )?;
-
-        if row_existed_before {
-            updated += 1;
-        } else {
-            inserted += 1;
-        }
-    }
-    tx.commit()?;
-    Ok((inserted, updated))
-}
-
-/// Parses an export blob from any reader. Extracted from `import_npcs` so the
-/// JSON-validation contract (the `"invalid JSON input"` context on malformed
-/// bytes) is unit-testable without a live stdin (TD-025).
-fn parse_import_blob(mut reader: impl Read) -> Result<ExportBlob> {
-    let mut input = String::new();
-    reader.read_to_string(&mut input)?;
-    serde_json::from_str(&input).context("invalid JSON input")
-}
-
-fn import_npcs(conn: &Connection) -> Result<()> {
-    let blob = parse_import_blob(io::stdin().lock())?;
-    let (inserted, updated) = import_npcs_inner(conn, blob.npcs)?;
-    println!(
-        "Imported NPCs from stdin: inserted {inserted}, updated {updated} (household_id and other non-export columns preserved on updates)"
-    );
-    Ok(())
-}
-
-fn family_tree(conn: &Connection, npc_id: i64) -> Result<()> {
-    // household_id is nullable in the schema (line 261). Fetch it as
-    // Option<i64> so a NULL value — possible via import or manual
-    // editing — doesn't surface as a misleading "NPC not found" error
-    // (#435). An NPC without a household still exists; we just have
-    // no tree to print.
-    let (household_id, target_name, target_age): (Option<i64>, String, i64) = conn
-        .query_row(
-            "SELECT household_id, name, age FROM npcs WHERE id = ?",
-            params![npc_id],
-            |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?)),
-        )
-        .optional()?
-        .context("NPC not found")?;
-
-    let Some(household_id) = household_id else {
-        println!("Family tree for {target_name}: no household assigned");
-        return Ok(());
-    };
-
-    println!("Family tree for {target_name} (household #{household_id})");
-    let mut stmt = conn
-        .prepare("SELECT id, name, age FROM npcs WHERE household_id = ? ORDER BY age DESC, id")?;
-    let rows = stmt.query_map(params![household_id], |r| {
-        Ok((
-            r.get::<_, i64>(0)?,
-            r.get::<_, String>(1)?,
-            r.get::<_, i64>(2)?,
-        ))
-    })?;
-
-    for row in rows {
-        let (id, name, age) = row?;
-        let relation = if id == npc_id {
-            "self"
-        } else if age >= target_age + 16 {
-            "possible elder"
-        } else if age + 16 <= target_age {
-            "possible younger"
-        } else {
-            "peer"
-        };
-        println!("- {id}: {name}, age {age} ({relation})");
-    }
-    Ok(())
-}
-
-fn relationships(conn: &Connection, npc_id: i64) -> Result<()> {
-    let exists: Option<String> = conn
-        .query_row("SELECT name FROM npcs WHERE id = ?", params![npc_id], |r| {
-            r.get(0)
-        })
-        .optional()?;
-    let name = exists.context("NPC not found")?;
-
-    println!("Relationships for {name} ({npc_id})");
-    let mut stmt = conn.prepare(
-        "
-        SELECT r.to_npc_id, n.name, r.kind, r.strength
-        FROM npc_relationships r
-        JOIN npcs n ON n.id = r.to_npc_id
-        WHERE r.from_npc_id = ?
-        ORDER BY r.strength DESC
-    ",
-    )?;
-    let rows = stmt.query_map(params![npc_id], |r| {
-        Ok((
-            r.get::<_, i64>(0)?,
-            r.get::<_, String>(1)?,
-            r.get::<_, String>(2)?,
-            r.get::<_, f64>(3)?,
-        ))
-    })?;
-
-    for row in rows {
-        let (target_id, target_name, kind, strength) = row?;
-        println!("- {target_id}: {target_name} [{kind}] {strength:.2}");
-    }
-    Ok(())
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn generated_conn(parish: &str, pop: u32, seed: u64) -> Connection {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world generation should work");
-        generate_parish(&conn, parish, pop, Some(seed)).expect("parish generation should work");
-        conn
-    }
-
-    fn assert_validation_failed(result: Result<()>) {
-        assert!(result.is_err(), "validation should fail");
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("validation failed"),
-            "validation error should use the aggregate failure message"
-        );
-    }
-
-    // ── weighted_occupation (TD-019) ─────────────────────────────────────────
-
-    #[test]
-    fn weighted_occupation_only_returns_known_occupations() {
-        // The weights sum to 100 so the `"Other"` fallback at the end of the
-        // loop is unreachable in practice; this RNG-driven sweep locks the
-        // distribution down — every draw must belong to OCCUPATIONS.
-        use std::collections::HashSet;
-        let known: HashSet<&str> = OCCUPATIONS.iter().map(|(occ, _)| *occ).collect();
-        let mut rng = StdRng::seed_from_u64(0xC0FFEE);
-        let mut seen: HashSet<&str> = HashSet::new();
-        for _ in 0..10_000 {
-            let occ = weighted_occupation(&mut rng);
-            assert!(
-                known.contains(occ),
-                "weighted_occupation returned an occupation outside the table: {occ}"
-            );
-            seen.insert(occ);
-        }
-        // Over 10k draws the high-weight occupations should all appear, proving
-        // the helper actually samples the table rather than always returning one.
-        assert!(
-            seen.contains("Tenant Farmer") && seen.contains("Laborer"),
-            "common occupations should be observed across 10k draws: {seen:?}"
-        );
-    }
-
-    // ── escape_like (TD-020) ─────────────────────────────────────────────────
-
-    #[test]
-    fn escape_like_escapes_backslash_percent_and_underscore() {
-        // Pure string-only coverage with no DB round-trip. The backslash branch
-        // (`\` -> `\\`) was previously only exercised indirectly via SQLite.
-        assert_eq!(escape_like("\\"), "\\\\");
-        assert_eq!(escape_like("%"), "\\%");
-        assert_eq!(escape_like("_"), "\\_");
-        // Backslash must be escaped first so the escapes of % and _ are not
-        // themselves re-escaped: `a\%_b` -> `a\\\%\_b`.
-        assert_eq!(escape_like("a\\%_b"), "a\\\\\\%\\_b");
-        // Plain text is left untouched.
-        assert_eq!(escape_like("Bridget"), "Bridget");
-    }
-
-    // ── DataTier round-trip (TD-021) ─────────────────────────────────────────
-
-    #[test]
-    fn data_tier_as_i64_from_i64_round_trip() {
-        for (tier, label) in [
-            (DataTier::Sketched, "Sketched"),
-            (DataTier::Elaborated, "Elaborated"),
-            (DataTier::Authored, "Authored"),
-        ] {
-            assert_eq!(DataTier::from_i64(tier.as_i64()), label);
-        }
-        // Explicit numeric contract.
-        assert_eq!(DataTier::Sketched.as_i64(), 0);
-        assert_eq!(DataTier::Elaborated.as_i64(), 1);
-        assert_eq!(DataTier::Authored.as_i64(), 2);
-    }
-
-    #[test]
-    fn data_tier_from_i64_out_of_range_is_unknown() {
-        // The "Unknown" fallback defends against schema drift / corrupt rows.
-        assert_eq!(DataTier::from_i64(-1), "Unknown");
-        assert_eq!(DataTier::from_i64(3), "Unknown");
-        assert_eq!(DataTier::from_i64(i64::MAX), "Unknown");
-    }
-
-    // ── import JSON-error path (TD-025) ──────────────────────────────────────
-
-    #[test]
-    fn parse_import_blob_rejects_malformed_json() {
-        use std::io::Cursor;
-        let err = parse_import_blob(Cursor::new(b"{ this is not json".to_vec()))
-            .expect_err("malformed JSON must error");
-        assert!(
-            err.to_string().contains("invalid JSON input"),
-            "JSON-validation context must bubble: {err}"
-        );
-    }
-
-    #[test]
-    fn parse_import_blob_accepts_valid_blob() {
-        use std::io::Cursor;
-        let json = br#"{"npcs":[{"id":1,"name":"Bridget","age":40,"parish":"kilteevan","occupation":"Servant","data_tier":0,"mood":null,"personality":null}]}"#;
-        let blob = parse_import_blob(Cursor::new(json.to_vec())).expect("valid blob parses");
-        assert_eq!(blob.npcs.len(), 1);
-        assert_eq!(blob.npcs[0].name, "Bridget");
-        // The `sex` field defaults when absent (legacy blob compatibility).
-        assert_eq!(blob.npcs[0].sex, "unknown");
-    }
-
-    #[test]
-    fn test_generate_world_rejects_empty_counties() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        let result = generate_world(&conn, &[]);
-        assert!(
-            result.is_err(),
-            "generate_world with empty counties must fail"
-        );
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("--counties is required"),
-            "error must mention --counties"
-        );
-    }
-
-    /// TD-023: `generate_parish` against a county-less DB must return a clear
-    /// `Err` (no panic, no silently auto-created `'roscommon'` county). A
-    /// read-only or constrained DB previously hit `.expect()` and aborted the
-    /// whole process; now it surfaces a recoverable error pointing at
-    /// `generate-world`.
-    #[test]
-    fn test_generate_parish_without_county_errors_not_panics() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        // No generate_world call → counties table is empty.
-        let result = generate_parish(&conn, "Kiltoom", 10, Some(1));
-        assert!(
-            result.is_err(),
-            "generate_parish with no county must return Err, not panic or auto-create one"
-        );
-        assert!(
-            result.unwrap_err().to_string().contains("generate-world"),
-            "error must direct the user to run generate-world first"
-        );
-        // And no county was silently invented on the user's behalf.
-        let county_count: i64 = conn
-            .query_row("SELECT COUNT(*) FROM counties", [], |r| r.get(0))
-            .expect("count query should succeed");
-        assert_eq!(
-            county_count, 0,
-            "generate_parish must not auto-create a county (TD-023)"
-        );
-    }
-
-    #[test]
-    fn test_schema_bootstrap_and_generation() {
-        let conn = generated_conn("Kiltoom", 30, 1);
-
-        let count: i64 = conn
-            .query_row("SELECT COUNT(*) FROM npcs", [], |r| r.get(0))
-            .expect("count query should succeed");
-        assert!(count > 0);
-    }
-
-    #[test]
-    fn test_promote_sets_personality() {
-        let conn = generated_conn("Kiltoom", 20, 2);
-
-        let npc_id: i64 = conn
-            .query_row("SELECT id FROM npcs ORDER BY id LIMIT 1", [], |r| r.get(0))
-            .expect("must have one NPC");
-        promote_npc(&conn, npc_id).expect("promotion should succeed");
-
-        let (tier, personality): (i64, Option<String>) = conn
-            .query_row(
-                "SELECT data_tier, personality FROM npcs WHERE id = ?",
-                params![npc_id],
-                |r| Ok((r.get(0)?, r.get(1)?)),
-            )
-            .expect("should read promoted NPC");
-        assert_eq!(tier, 1);
-        assert!(personality.is_some());
-    }
-
-    #[test]
-    fn test_promote_rejects_missing_target() {
-        let conn = generated_conn("Kiltoom", 20, 2);
-
-        let result = promote_npc(&conn, 9_999_999);
-        assert!(result.is_err(), "promoting a missing NPC should fail");
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("NPC 9999999 not found"),
-            "error should name the missing NPC"
-        );
-    }
-
-    #[test]
-    fn test_validate_detects_missing_personality() {
-        let conn = generated_conn("Kiltoom", 20, 3);
-
-        conn.execute(
-            "UPDATE npcs SET data_tier = 1, personality = '' WHERE id = (SELECT id FROM npcs LIMIT 1)",
-            [],
-        )
-        .expect("update should succeed");
-
-        assert_validation_failed(validate_db(&conn, None, true));
-    }
-
-    #[test]
-    fn test_validate_detects_missing_household() {
-        let conn = generated_conn("Kiltoom", 20, 4);
-
-        conn.execute(
-            "UPDATE npcs SET household_id = NULL WHERE id = (SELECT id FROM npcs LIMIT 1)",
-            [],
-        )
-        .expect("update should succeed");
-
-        assert_validation_failed(validate_db(&conn, None, true));
-    }
-
-    #[test]
-    fn test_validate_detects_invalid_age() {
-        let conn = generated_conn("Kiltoom", 20, 5);
-
-        conn.execute(
-            "UPDATE npcs SET age = 111 WHERE id = (SELECT id FROM npcs LIMIT 1)",
-            [],
-        )
-        .expect("update should succeed");
-
-        assert_validation_failed(validate_db(&conn, None, true));
-    }
-
-    #[test]
-    fn test_validate_detects_broken_relationship() {
-        let conn = generated_conn("Kiltoom", 20, 6);
-        let npc_id: i64 = conn
-            .query_row("SELECT id FROM npcs ORDER BY id LIMIT 1", [], |r| r.get(0))
-            .expect("must have an NPC");
-
-        conn.execute_batch("PRAGMA foreign_keys = OFF")
-            .expect("foreign keys should be configurable");
-        conn.execute(
-            "INSERT INTO npc_relationships(from_npc_id, to_npc_id, kind, strength) VALUES (?, 9999999, 'Acquaintance', 0.5)",
-            params![npc_id],
-        )
-        .expect("insert broken relationship");
-        conn.execute_batch("PRAGMA foreign_keys = ON")
-            .expect("foreign keys should be configurable");
-
-        assert_validation_failed(validate_db(&conn, None, true));
-    }
-
-    #[test]
-    fn test_validate_rejects_parish_and_all_together() {
-        let conn = generated_conn("Kiltoom", 20, 7);
-
-        let result = validate_db(&conn, Some("Kiltoom".to_string()), true);
-        assert!(result.is_err(), "validate should reject ambiguous scope");
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("choose either --parish or --all"),
-            "error should explain the mutually exclusive scope flags"
-        );
-    }
-
-    #[test]
-    fn test_generate_parish_same_seed_is_deterministic() {
-        fn npc_signature(conn: &Connection) -> Vec<(String, String, i64, i64, String, String)> {
-            let mut stmt = conn
-                .prepare(
-                    "
-                    SELECT n.name, n.sex, n.birth_year, n.age, h.name, n.occupation
-                    FROM npcs n
-                    JOIN households h ON h.id = n.household_id
-                    ORDER BY n.id
-                    ",
-                )
-                .expect("prepare NPC signature query");
-            stmt.query_map([], |r| {
-                Ok((
-                    r.get::<_, String>(0)?,
-                    r.get::<_, String>(1)?,
-                    r.get::<_, i64>(2)?,
-                    r.get::<_, i64>(3)?,
-                    r.get::<_, String>(4)?,
-                    r.get::<_, String>(5)?,
-                ))
-            })
-            .expect("query NPC signature")
-            .collect::<rusqlite::Result<Vec<_>>>()
-            .expect("collect NPC signature")
-        }
-
-        fn relationship_signature(conn: &Connection) -> Vec<(i64, i64, String, String)> {
-            let mut stmt = conn
-                .prepare(
-                    "
-                    SELECT from_npc_id, to_npc_id, kind, printf('%.6f', strength)
-                    FROM npc_relationships
-                    ORDER BY from_npc_id, to_npc_id, kind
-                    ",
-                )
-                .expect("prepare relationship signature query");
-            stmt.query_map([], |r| {
-                Ok((
-                    r.get::<_, i64>(0)?,
-                    r.get::<_, i64>(1)?,
-                    r.get::<_, String>(2)?,
-                    r.get::<_, String>(3)?,
-                ))
-            })
-            .expect("query relationship signature")
-            .collect::<rusqlite::Result<Vec<_>>>()
-            .expect("collect relationship signature")
-        }
-
-        let left = generated_conn("Kiltoom", 30, 8);
-        let right = generated_conn("Kiltoom", 30, 8);
-
-        assert_eq!(npc_signature(&left), npc_signature(&right));
-        assert_eq!(
-            relationship_signature(&left),
-            relationship_signature(&right)
-        );
-    }
-
-    // ── #436 import preserves non-export columns + sex round-trips ──────────
-
-    /// Seeds one NPC with a known sex and household_id, then simulates
-    /// the import path on a blob that represents re-importing that NPC
-    /// with updated personality. household_id must survive untouched and
-    /// sex must come from the blob (not hard-coded 'unknown').
-    #[test]
-    fn test_import_preserves_household_and_restores_sex() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world generation should work");
-
-        let parish_id: i64 = conn
-            .query_row("SELECT id FROM parishes LIMIT 1", [], |r| r.get(0))
-            .ok()
-            .unwrap_or_else(|| {
-                conn.execute(
-                    "INSERT INTO parishes(county_id, name) VALUES ((SELECT id FROM counties LIMIT 1), 'Testshire')",
-                    [],
-                )
-                .unwrap();
-                conn.last_insert_rowid()
-            });
-
-        // Insert a household so we have a non-NULL household_id to preserve.
-        conn.execute(
-            "INSERT INTO households(parish_id, name) VALUES (?, 'Darcy')",
-            params![parish_id],
-        )
-        .unwrap();
-        let household_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "INSERT INTO npcs(id, name, sex, birth_year, age, parish_id, household_id, occupation, data_tier, mood)\n             VALUES (42, 'Pádraig Darcy', 'male', 1762, 58, ?, ?, 'Publican', 1, 'content')",
-            params![parish_id, household_id],
-        )
-        .unwrap();
-
-        // Build an import blob that updates personality but carries the
-        // NPC's existing id. `sex` is present (no longer hard-coded).
-        // parish is reused so we don't hit unrelated lookup paths.
-        let parish_name: String = conn
-            .query_row(
-                "SELECT name FROM parishes WHERE id = ?",
-                params![parish_id],
-                |r| r.get(0),
-            )
-            .unwrap();
-        let blob = ExportBlob {
-            npcs: vec![ExportNpc {
-                id: 42,
-                name: "Pádraig Darcy".to_string(),
-                sex: "male".to_string(),
-                age: 58,
-                parish: parish_name,
-                occupation: "Publican".to_string(),
-                data_tier: 1,
-                mood: Some("content".to_string()),
-                personality: Some("Warm-hearted publican.".to_string()),
-            }],
-        };
-
-        // Use the shared import helper instead of duplicating the UPSERT SQL.
-        import_npcs_inner(&conn, blob.npcs).unwrap();
-
-        // household_id must still be set (would be NULL if INSERT OR
-        // REPLACE were used — that was the #436 regression).
-        let (hh, sex, personality): (Option<i64>, String, Option<String>) = conn
-            .query_row(
-                "SELECT household_id, sex, personality FROM npcs WHERE id = 42",
-                [],
-                |r| Ok((r.get(0)?, r.get(1)?, r.get(2)?)),
-            )
-            .unwrap();
-        assert_eq!(hh, Some(household_id), "household_id must survive import");
-        assert_eq!(sex, "male", "sex must come from the blob, not 'unknown'");
-        assert_eq!(
-            personality.as_deref(),
-            Some("Warm-hearted publican."),
-            "personality must update on import",
-        );
-    }
-
-    /// A blob serialized *before* #436 (no `sex` field) must still
-    /// deserialize cleanly, defaulting to "unknown" — so we don't
-    /// break users with saved export files from earlier versions.
-    #[test]
-    fn test_export_blob_deserializes_legacy_missing_sex() {
-        let legacy = r#"{
-            "npcs": [{
-                "id": 1,
-                "name": "Legacy Mary",
-                "age": 40,
-                "parish": "Kiltoom",
-                "occupation": "Servant",
-                "data_tier": 0,
-                "mood": null,
-                "personality": null
-            }]
-        }"#;
-        let blob: ExportBlob = serde_json::from_str(legacy).expect("legacy blob should parse");
-        assert_eq!(blob.npcs.len(), 1);
-        assert_eq!(blob.npcs[0].sex, "unknown");
-    }
-
-    // ── #435 family_tree gracefully handles NULL household_id ───────────────
-
-    /// An NPC with NULL household_id must not blow up family_tree
-    /// with a misleading "NPC not found" error. The NPC exists — we
-    /// just have no household to walk.
-    #[test]
-    fn test_family_tree_handles_null_household() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world generation should work");
-
-        // Insert an NPC directly with NULL household_id — import or
-        // manual editing can produce this in the wild.
-        let parish_id: i64 = conn
-            .query_row("SELECT id FROM parishes LIMIT 1", [], |r| r.get(0))
-            .ok()
-            .unwrap_or_else(|| {
-                conn.execute(
-                    "INSERT INTO parishes(county_id, name) VALUES ((SELECT id FROM counties LIMIT 1), 'Testshire')",
-                    [],
-                )
-                .unwrap();
-                conn.last_insert_rowid()
-            });
-        conn.execute(
-            "INSERT INTO npcs(name, sex, birth_year, age, parish_id, household_id, occupation, data_tier, mood)\n             VALUES ('Orphan', 'female', 1790, 30, ?, NULL, 'Other', 0, 'neutral')",
-            params![parish_id],
-        )
-        .expect("insert orphan NPC");
-        let orphan_id = conn.last_insert_rowid();
-
-        // The call must succeed (return Ok) rather than surfacing a
-        // confusing "NPC not found" error.
-        let result = family_tree(&conn, orphan_id);
-        assert!(
-            result.is_ok(),
-            "family_tree on NULL-household NPC should succeed, got: {:?}",
-            result.err()
-        );
-
-        // And a non-existent NPC id still reports "NPC not found"
-        // (regression-guard: we didn't break that path).
-        let missing = family_tree(&conn, 9_999_999);
-        assert!(missing.is_err());
-        assert!(
-            missing.unwrap_err().to_string().contains("NPC not found"),
-            "missing NPC should still surface 'NPC not found'"
-        );
-    }
-
-    // ── #607 search_npcs escapes LIKE wildcard metacharacters ──────────────
-
-    fn setup_search_db() -> Connection {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-
-        let county_id: i64 = conn
-            .query_row(
-                "INSERT INTO counties(name) VALUES ('Roscommon') RETURNING id",
-                [],
-                |r| r.get(0),
-            )
-            .unwrap_or_else(|_| {
-                conn.execute("INSERT INTO counties(name) VALUES ('Roscommon')", [])
-                    .unwrap();
-                conn.last_insert_rowid()
-            });
-
-        conn.execute(
-            "INSERT INTO parishes(county_id, name) VALUES (?, 'Kiltoom')",
-            params![county_id],
-        )
-        .unwrap();
-        let parish_id = conn.last_insert_rowid();
-
-        for (name, occ) in &[
-            ("100%_off", "Merchant"),
-            ("Alice the Smith", "Blacksmith"),
-            ("O_Brien", "Farmer"),
-        ] {
-            conn.execute(
-                "INSERT INTO npcs(name, sex, birth_year, age, parish_id, occupation, data_tier, mood) \
-                 VALUES (?, 'unknown', 1800, 20, ?, ?, 0, 'neutral')",
-                params![name, parish_id, occ],
-            )
-            .unwrap();
-        }
-
-        conn
-    }
-
-    fn search_names(conn: &Connection, query: &str) -> Vec<String> {
-        let like = format!("%{}%", escape_like(query));
-        let mut stmt = conn
-            .prepare("SELECT n.name FROM npcs n WHERE n.name LIKE ? ESCAPE '\\' ORDER BY n.name")
-            .unwrap();
-        stmt.query_map(params![like], |r| r.get::<_, String>(0))
-            .unwrap()
-            .map(|r| r.unwrap())
-            .collect()
-    }
-
-    #[test]
-    fn test_search_wildcard_chars_match_literal_only() {
-        let conn = setup_search_db();
-        let results = search_names(&conn, "100%_off");
-        assert_eq!(results, vec!["100%_off".to_string()]);
-    }
-
-    #[test]
-    fn test_search_normal_query_still_matches() {
-        let conn = setup_search_db();
-        let results = search_names(&conn, "alice");
-        assert_eq!(results, vec!["Alice the Smith".to_string()]);
-    }
-
-    // ── #592 SQL injection guard ─────────────────────────────────────────────
-
-    /// A parish name containing SQL meta-characters must not cause query
-    /// errors or allow arbitrary SQL execution.  The parameterized query
-    /// should treat the entire string as a literal value; because no parish
-    /// with that name exists, all counts come back as zero and validation
-    /// passes (empty DB for the injected "parish").
-    #[test]
-    fn validate_db_parish_filter_rejects_sqli_payload() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-
-        // SQL injection payloads: these would break string-interpolated queries.
-        let payloads = [
-            "Kiltoom'); DROP TABLE npcs; --",
-            "' OR '1'='1",
-            "x' UNION SELECT 0,0,0,0,0 --",
-            "Kil'toom",
-        ];
-
-        for payload in payloads {
-            // Must not return a rusqlite error — the value is bound, not
-            // interpolated, so the query is syntactically valid regardless.
-            let result = validate_db(&conn, Some(payload.to_string()), false);
-            assert!(
-                result.is_ok(),
-                "validate_db should not error on payload {payload:?}, got: {:?}",
-                result.err()
-            );
-        }
-
-        // Sanity check: the npcs table still exists (no DROP TABLE succeeded).
-        let count: i64 = conn
-            .query_row("SELECT COUNT(*) FROM npcs", [], |r| r.get(0))
-            .expect("npcs table must still exist after injection attempts");
-        assert_eq!(count, 0, "no NPCs were inserted so count must be 0");
-    }
-
-    // ── TD-002 list_npcs tests ────────────────────────────────────────────────
-
-    fn setup_list_db() -> Connection {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world gen");
-        generate_parish(&conn, "Kiltoom", 20, Some(1)).expect("parish gen");
-        conn
-    }
-
-    #[test]
-    fn test_list_npcs_unfiltered() {
-        let conn = setup_list_db();
-        let result = list_npcs(&conn, None, None, None, 100);
-        assert!(result.is_ok(), "unfiltered list should succeed");
-    }
-
-    #[test]
-    fn test_list_npcs_parish_filter() {
-        let conn = setup_list_db();
-        let result = list_npcs(&conn, Some("Kiltoom"), None, None, 100);
-        assert!(result.is_ok(), "parish-filtered list should succeed");
-    }
-
-    #[test]
-    fn test_list_npcs_occupation_filter() {
-        let conn = setup_list_db();
-        let result = list_npcs(&conn, None, Some("Laborer"), None, 100);
-        assert!(result.is_ok(), "occupation-filtered list should succeed");
-    }
-
-    #[test]
-    fn test_list_npcs_tier_filter() {
-        let conn = setup_list_db();
-        let result = list_npcs(&conn, None, None, Some(DataTier::Sketched), 100);
-        assert!(result.is_ok(), "tier-filtered list should succeed");
-    }
-
-    #[test]
-    fn test_list_npcs_all_filters() {
-        let conn = setup_list_db();
-        let result = list_npcs(
-            &conn,
-            Some("Kiltoom"),
-            Some("Tenant Farmer"),
-            Some(DataTier::Sketched),
-            50,
-        );
-        assert!(result.is_ok(), "all-filters list should succeed");
-    }
-
-    #[test]
-    fn test_list_npcs_empty_result() {
-        let conn = setup_list_db();
-        let result = list_npcs(&conn, Some("Nonexistent"), None, None, 10);
-        assert!(result.is_ok(), "list with no matches should succeed");
-    }
-
-    // ── TD-003 show_npc tests ────────────────────────────────────────────────
-
-    #[test]
-    fn test_show_npc_found() {
-        let conn = setup_list_db();
-        let npc_id: i64 = conn
-            .query_row("SELECT id FROM npcs ORDER BY id LIMIT 1", [], |r| r.get(0))
-            .expect("must have an NPC");
-        let result = show_npc(&conn, npc_id);
-        assert!(result.is_ok(), "show_npc for existing NPC should succeed");
-    }
-
-    #[test]
-    fn test_show_npc_not_found() {
-        let conn = setup_list_db();
-        let result = show_npc(&conn, 9_999_999);
-        assert!(result.is_err(), "show_npc for missing NPC should fail");
-        assert!(
-            result.unwrap_err().to_string().contains("not found"),
-            "error must mention 'not found'"
-        );
-    }
-
-    // ── TD-004 edit_npc tests ────────────────────────────────────────────────
-
-    fn setup_edit_db() -> (Connection, i64) {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world gen");
-        generate_parish(&conn, "Kiltoom", 10, Some(2)).expect("parish gen");
-        let npc_id: i64 = conn
-            .query_row("SELECT id FROM npcs ORDER BY id LIMIT 1", [], |r| r.get(0))
-            .expect("must have an NPC");
-        (conn, npc_id)
-    }
-
-    #[test]
-    fn test_edit_npc_mood_only() {
-        let (conn, npc_id) = setup_edit_db();
-        let result = edit_npc(&conn, npc_id, Some("happy".to_string()), None);
-        assert!(result.is_ok(), "edit mood should succeed");
-        let mood: Option<String> = conn
-            .query_row("SELECT mood FROM npcs WHERE id = ?", params![npc_id], |r| {
-                r.get(0)
-            })
-            .unwrap();
-        assert_eq!(mood.as_deref(), Some("happy"));
-    }
-
-    #[test]
-    fn test_edit_npc_occupation_only() {
-        let (conn, npc_id) = setup_edit_db();
-        let result = edit_npc(&conn, npc_id, None, Some("Publican".to_string()));
-        assert!(result.is_ok(), "edit occupation should succeed");
-        let occ: String = conn
-            .query_row(
-                "SELECT occupation FROM npcs WHERE id = ?",
-                params![npc_id],
-                |r| r.get(0),
-            )
-            .unwrap();
-        assert_eq!(occ, "Publican");
-    }
-
-    #[test]
-    fn test_edit_npc_both() {
-        let (conn, npc_id) = setup_edit_db();
-        let result = edit_npc(
-            &conn,
-            npc_id,
-            Some("sad".to_string()),
-            Some("Laborer".to_string()),
-        );
-        assert!(result.is_ok(), "edit both should succeed");
-        let (mood, occ): (Option<String>, String) = conn
-            .query_row(
-                "SELECT mood, occupation FROM npcs WHERE id = ?",
-                params![npc_id],
-                |r| Ok((r.get(0)?, r.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(mood.as_deref(), Some("sad"));
-        assert_eq!(occ, "Laborer");
-    }
-
-    #[test]
-    fn test_edit_npc_no_changes() {
-        let (conn, npc_id) = setup_edit_db();
-        let result = edit_npc(&conn, npc_id, None, None);
-        assert!(result.is_err(), "edit with no changes should fail");
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("provide at least one change"),
-            "error must mention providing at least one change"
-        );
-    }
-
-    // ── TD-005 elaborate_parish tests ────────────────────────────────────────
-
-    #[test]
-    fn test_elaborate_parish_basic() {
-        let conn = setup_list_db();
-        let result = elaborate_parish(&conn, "Kiltoom", 5);
-        assert!(result.is_ok(), "elaborate_parish should succeed");
-        let elaborated: i64 = conn
-            .query_row("SELECT COUNT(*) FROM npcs WHERE data_tier >= 1", [], |r| {
-                r.get(0)
-            })
-            .unwrap();
-        assert_eq!(elaborated, 5, "should have elaborated 5 NPCs");
-    }
-
-    #[test]
-    fn test_elaborate_parish_empty_result() {
-        let conn = setup_list_db();
-        let result = elaborate_parish(&conn, "Nonexistent", 10);
-        assert!(
-            result.is_ok(),
-            "elaborate_parish with no matches should succeed"
-        );
-    }
-
-    #[test]
-    fn test_elaborate_parish_limit_zero() {
-        let conn = setup_list_db();
-        let result = elaborate_parish(&conn, "Kiltoom", 0);
-        assert!(
-            result.is_ok(),
-            "elaborate_parish with limit 0 should succeed"
-        );
-    }
-
-    // ── TD-008 relationships tests ───────────────────────────────────────────
-
-    #[test]
-    fn test_relationships_npc_found() {
-        let conn = setup_list_db();
-        let npc_id: i64 = conn
-            .query_row("SELECT id FROM npcs ORDER BY id LIMIT 1", [], |r| r.get(0))
-            .expect("must have an NPC");
-        let result = relationships(&conn, npc_id);
-        assert!(
-            result.is_ok(),
-            "relationships for existing NPC should succeed"
-        );
-    }
-
-    #[test]
-    fn test_relationships_npc_not_found() {
-        let conn = setup_list_db();
-        let result = relationships(&conn, 9_999_999);
-        assert!(result.is_err(), "relationships for missing NPC should fail");
-        assert!(
-            result.unwrap_err().to_string().contains("NPC not found"),
-            "error must mention 'NPC not found'"
-        );
-    }
-
-    // ── TD-006 stats tests ───────────────────────────────────────────────────
-
-    #[test]
-    fn test_stats_with_data() {
-        let conn = setup_list_db();
-        let result = stats(&conn);
-        assert!(result.is_ok(), "stats with data should succeed");
-    }
-
-    #[test]
-    fn test_stats_empty_db() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        let result = stats(&conn);
-        assert!(result.is_ok(), "stats on empty DB should succeed");
-    }
-
-    // ── TD-007 export_npcs tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_export_npcs_unfiltered() {
-        let conn = setup_list_db();
-        let result = export_npcs(&conn, None);
-        assert!(result.is_ok(), "unfiltered export should succeed");
-    }
-
-    #[test]
-    fn test_export_npcs_parish_filtered() {
-        let conn = setup_list_db();
-        let result = export_npcs(&conn, Some("Kiltoom"));
-        assert!(result.is_ok(), "parish-filtered export should succeed");
-    }
-
-    #[test]
-    fn test_export_npcs_empty_result() {
-        let conn = setup_list_db();
-        let result = export_npcs(&conn, Some("Nonexistent"));
-        assert!(result.is_ok(), "export with no matches should succeed");
-    }
-
-    // ── TD-009 search_npcs tests ─────────────────────────────────────────────
-
-    #[test]
-    fn test_search_npcs_matches() {
-        let conn = setup_search_db();
-        let result = search_npcs(&conn, "alice", 10);
-        assert!(result.is_ok(), "search for existing name should succeed");
-    }
-
-    #[test]
-    fn test_search_npcs_no_matches() {
-        let conn = setup_search_db();
-        let result = search_npcs(&conn, "zzzznotfound", 10);
-        assert!(result.is_ok(), "search for missing name should succeed");
-    }
-
-    #[test]
-    fn test_search_npcs_limit_zero() {
-        let conn = setup_search_db();
-        let result = search_npcs(&conn, "alice", 0);
-        assert!(result.is_ok(), "search with limit 0 should succeed");
-    }
-
-    /// A legitimate parish name must still filter correctly — the fix must not
-    /// break the happy path.
-    #[test]
-    fn validate_db_parish_filter_works_for_valid_name() {
-        let conn = Connection::open_in_memory().expect("in-memory SQLite should open");
-        ensure_schema(&conn).expect("schema should initialize");
-        generate_world(&conn, &["roscommon".to_string()]).expect("world generation should work");
-        generate_parish(&conn, "Kiltoom", 5, Some(4)).expect("parish generation should work");
-
-        // All generated NPCs should have a household and valid age, so
-        // validate_db with the real parish name should pass.
-        let result = validate_db(&conn, Some("Kiltoom".to_string()), false);
-        assert!(
-            result.is_ok(),
-            "validate_db with a real parish name must pass, got: {:?}",
-            result.err()
-        );
-    }
-
-    // ── resolve_default_db (TD-024 / Rule 9) ────────────────────────────────
-
-    /// Holds the previous value of an env var and restores it on drop.
-    /// Env mutations must be done inside a single-threaded guard because
-    /// Cargo runs tests in the same process concurrently.
-    struct EnvGuard {
-        key: &'static str,
-        prev: Option<std::ffi::OsString>,
-    }
-    impl EnvGuard {
-        fn capture(key: &'static str) -> Self {
-            EnvGuard {
-                key,
-                prev: std::env::var_os(key),
-            }
-        }
-    }
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            // SAFETY: gated by DB_RESOLVE_LOCK.
-            match self.prev.take() {
-                Some(v) => unsafe { std::env::set_var(self.key, v) },
-                None => unsafe { std::env::remove_var(self.key) },
-            }
-        }
-    }
-
-    fn db_resolve_lock() -> std::sync::MutexGuard<'static, ()> {
-        static LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-        LOCK.lock().unwrap_or_else(|e| e.into_inner())
-    }
-
-    #[test]
-    fn resolve_default_db_npc_tool_env_wins() {
-        let _gate = db_resolve_lock();
-        let _g1 = EnvGuard::capture(NPC_TOOL_DB_ENV);
-        let _g2 = EnvGuard::capture("PARISH_DATA_DIR");
-
-        // SAFETY: gated.
-        unsafe {
-            std::env::set_var(NPC_TOOL_DB_ENV, "/tmp/override.db");
-            std::env::remove_var("PARISH_DATA_DIR");
-        }
-
-        let resolved = resolve_default_db();
-        assert_eq!(
-            resolved,
-            PathBuf::from("/tmp/override.db"),
-            "PARISH_NPC_TOOL_DB must be used verbatim"
-        );
-    }
-
-    #[test]
-    fn resolve_default_db_parish_data_dir_second_priority() {
-        let _gate = db_resolve_lock();
-        let _g1 = EnvGuard::capture(NPC_TOOL_DB_ENV);
-        let _g2 = EnvGuard::capture("PARISH_DATA_DIR");
-
-        // SAFETY: gated.
-        unsafe {
-            std::env::remove_var(NPC_TOOL_DB_ENV);
-            std::env::set_var("PARISH_DATA_DIR", "/tmp/mydata");
-        }
-
-        let resolved = resolve_default_db();
-        assert_eq!(
-            resolved,
-            PathBuf::from("/tmp/mydata").join(DB_FILENAME),
-            "PARISH_DATA_DIR must yield <dir>/parish-world.db"
-        );
-    }
-
-    #[test]
-    fn resolve_default_db_empty_env_var_is_skipped() {
-        let _gate = db_resolve_lock();
-        let _g1 = EnvGuard::capture(NPC_TOOL_DB_ENV);
-        let _g2 = EnvGuard::capture("PARISH_DATA_DIR");
-
-        // SAFETY: gated.
-        unsafe {
-            std::env::set_var(NPC_TOOL_DB_ENV, "   ");
-            std::env::set_var("PARISH_DATA_DIR", "   ");
-        }
-
-        // Both trimmed-empty — should fall through to ancestor walk or cwd fallback.
-        // The result will be some path ending in `data/parish-world.db`; the key
-        // property is that neither blank env var was used.
-        let resolved = resolve_default_db();
-        assert!(
-            resolved.ends_with(PathBuf::from("data").join(DB_FILENAME)),
-            "blank env vars must be ignored; got: {}",
-            resolved.display()
-        );
-    }
-
-    #[test]
-    fn resolve_default_db_falls_back_to_data_subdir() {
-        let _gate = db_resolve_lock();
-        let _g1 = EnvGuard::capture(NPC_TOOL_DB_ENV);
-        let _g2 = EnvGuard::capture("PARISH_DATA_DIR");
-
-        // SAFETY: gated.
-        unsafe {
-            std::env::remove_var(NPC_TOOL_DB_ENV);
-            std::env::remove_var("PARISH_DATA_DIR");
-        }
-
-        // Whether we find Cargo.toml or fall all the way to the cwd fallback,
-        // the result must end with data/parish-world.db.
-        let resolved = resolve_default_db();
-        assert!(
-            resolved.ends_with(PathBuf::from("data").join(DB_FILENAME)),
-            "fallback must end with data/parish-world.db; got: {}",
-            resolved.display()
-        );
-        assert!(
-            resolved.is_absolute(),
-            "resolved path must be absolute (anchored); got: {}",
-            resolved.display()
-        );
-    }
-}
+mod tests;
